refactor(rag-knowledge): U2a — Service 系 3 Port + Adapter 新設、facade に委譲化 (#730)
Issue #730 の Unit 2a。Service 系 3 Port + RealAdapter を新設し、
`RAGKnowledgeService` を新 Adapter への委譲 facade に書き換える。
本 commit 段階では facade を残し、`rag_knowledge.py` 自体は削除しない
（cli.py / evaluation.py の呼び出し側切替・rag_knowledge.py 削除は U2 後続で対応）。

新 Port + Adapter:

- `src/rag/search/search_port.py`
  - `SearchPort` Protocol + `RealSearchAdapter` 実装
  - `retrieve` / `retrieve_raw_results` / `_retrieve_vector_only` / `_retrieve_hybrid`
- `src/rag/search/filters.py`
  - `build_where_clause` / `build_bm25_filters` 関数
- `src/rag/admin/source_management_port.py`
  - `SourceManagementPort` Protocol + `RealSourceManagementAdapter` 実装
  - `source_exists` / `get_full_page_text` / `delete_source` / `get_document`
  - `get_document` 関数（自由関数）も再 export
- `src/rag/admin/stats_port.py`
  - `StatsPort` Protocol + `RealStatsAdapter` 実装
  - `get_stats` / `list_recent`
  - `list_recent_sources` 関数（自由関数）も再 export

`rag_knowledge.py`:

- `RAGKnowledgeService` を新 Adapter への委譲 facade に書き換え
  - `__init__` で各 Adapter (`RealSearchAdapter` / `RealSourceManagementAdapter` /
    `RealStatsAdapter`) を構築
  - 公開メソッド（`retrieve` / `retrieve_raw_results` / `source_exists` /
    `get_full_page_text` / `delete_source` / `get_stats`）は内部 Adapter に
    委譲
  - `_ingest_crawled_page` は評価専用（Issue #739 で運用判断中）として facade に
    暫定保持
  - 互換用 proxy property `_hybrid_search_engine` を追加（テスト互換のため）
- LOC: 782 → 約 230

テスト追従:

- `tests/test_rag_knowledge.py`: caplog の logger 名を `rag.rag_knowledge` から
  `rag.search.search_port` に更新（Adapter 化に伴い logger module が変更）
- `tests/test_search_filters.py`: `RAGKnowledgeService._build_where_clause` /
  `_build_bm25_filters` の参照を `rag.search.filters.build_where_clause` /
  `build_bm25_filters` に更新

検証:

- ruff / mypy 0 error（src/rag 129 source files）
- pytest tests/test_rag_knowledge.py tests/test_rag_knowledge_hybrid.py
  tests/test_search_filters.py tests/test_get_document.py tests/test_list_recent.py
  118 件 pass
